+++ play/components/Constructor.vue
@@ -1,30 +1,32 @@
 <template>
-  <form class="material-field-constructor-form">
-    <a href="#">back</a>
-
-    <p>Input constructor</p>
+  <form class="example">
     <MaterialInput v-bind="childProperies">
       {{ text }}
     </MaterialInput>
 
     <p>Modify it!</p>
 
-    <div class="contructor-field">
+    <div class="example-field">
       <label>Text</label>
       <input type="text" v-model="text" />
     </div>
 
-    <div class="contructor-field">
+    <div class="example-field">
       <label>Placeholder</label>
       <input type="text" v-model="childProperies.placeholder" />
     </div>
 
-    <div class="contructor-field">
+    <div class="example-field">
+      <label>Value</label>
+      <input type="text" v-model="childProperies.value" />
+    </div>
+
+    <div class="example-field">
       <label>Disabled</label>
       <input type="checkbox" v-model="childProperies.disabled" />
     </div>
 
-    <div class="contructor-field">
+    <div class="example-field">
       <label>Read-only</label>
       <input type="checkbox" v-model="childProperies.readonly" />
     </div>
@@ -33,9 +35,9 @@
 </template>
 
 <script>
-  import MaterialInput from '../src/components/MaterialInput.vue'
+  import MaterialInput from '../../src/components/MaterialInput.vue'
 
-  module.exports = {
+  export default {
     name: 'material-model',
     data () {
       return {
@@ -43,6 +45,7 @@
         childProperies: {
           type: 'text',
           placeholder: null,
+          value: null,
           disabled: false,
           readonly: false
         }
@@ -53,15 +56,3 @@
     }
   }
 </script>
-
-<style lang="scss">
-  .material-field-constructor-form {
-    max-width: 300px;
-    display: block;
-    margin: 0 auto;
-
-    p {
-      padding: 30px 0;
-    }
-  }
-</style>