+++ play/components/InputTypes.vue
@@ -1,8 +1,5 @@
 <template>
-  <form class="material-field-type-form">
-    <a href="#">back</a>
-
-    <p>Input types</p>
+  <form class="example">
     <MaterialInput name="name" type="text">Text</MaterialInput>
     <MaterialInput name="name" type="password">Password</MaterialInput>
     <MaterialInput name="name" type="number">Number</MaterialInput>
@@ -13,24 +10,12 @@
 </template>
 
 <script>
-  import MaterialInput from '../src/components/MaterialInput.vue'
+  import MaterialInput from '../../src/components/MaterialInput.vue'
 
-  module.exports = {
+  export default {
     name: 'material-model',
     components: {
       MaterialInput
     }
   }
 </script>
-
-<style lang="scss">
-  .material-field-type-form {
-    max-width: 300px;
-    display: block;
-    margin: 0 auto;
-
-    p {
-      padding: 30px 0;
-    }
-  }
-</style>