+++ test/unit/karma.conf.js
@@ -65,7 +65,6 @@ module.exports = function (config) {
     frameworks: ['mocha', 'sinon-chai'],
     reporters: ['spec', 'coverage', 'coveralls'],
     files: [
-      '../../node_modules/babel-polyfill/dist/polyfill.js',
       './index.js'
     ],
     preprocessors: {