+++ test/unit/specs/MaterialInput.spec.js
@@ -8,6 +8,8 @@ describe('MaterialInput.vue', () => {
     component = getComponent(MaterialInput, {})
   })
 
+  // Inputs
+
   it('renders correct default input type', () => {
     expect(component.$el.querySelector('input').type)
       .to.equal('text')
@@ -19,11 +21,15 @@ describe('MaterialInput.vue', () => {
       .to.equal('number')
   })
 
+  // Values
+
   it('copies the value when provided', () => {
     const inst = getComponent(MaterialInput, {value: 'some'})
     expect(inst.value).to.equal('some')
     expect(inst.value).to.equal(inst.valueCopy)
   })
+
+  // Classes
 
   it('returns correct default computedClasses', () => {
     expect(component.computedClasses).to.deep.equal([])
@@ -51,10 +57,17 @@ describe('MaterialInput.vue', () => {
     ])
   })
 
-  it('sets valueCopy to be equal to passed value', () => {
-    const inst = getComponent(MaterialInput, {value: 'some'})
-    expect(inst.valueCopy).to.equal('some')
+  it('returns correct computedClasses with errors', () => {
+    const errors = getComponent(MaterialInput, {
+      errorMessages: ['Error']
+    })
+
+    expect(errors.computedClasses).to.deep.equal([
+      'material-input--has-errors'
+    ])
   })
+
+  // Validation
 
   it('sets valid state on valid input', () => {
     const inst = getComponent(MaterialInput, {
@@ -65,19 +78,13 @@ describe('MaterialInput.vue', () => {
     expect(input.validity.valid).to.equal(true)
   })
 
-  it('sets event listener when pattern prop is passed', () => {
+  it('sets invalid state on invalid input', () => {
     const inst = getComponent(MaterialInput, {
-      pattern: {
-        regex: /\./,
-        message: 'Dot is required'
-      },
-      required: false
+      required: true,
+      type: 'email'
     })
 
     const input = inst.$el.querySelector('input')
-    /* eslint-disable no-undef */
-    input.dispatchEvent(new Event('input'))
-
     expect(input.validity.valid).to.equal(false)
   })
 })